feat&fix: add label repo and modif another file

+++ models/card.go
@@ -7,17 +7,17 @@ import (
 )
 
 type Card struct {
-	InternalID int64 `json:"internal_id" db:"internal_id" gorm:"primaryKey;autoIncrement"`
-	PublicID   uuid.UUID `json:"public_id" db:"public_id"`
-	ListID int64 `json:"list_internal_id" db:"list_internal_id" gorm:"column:list_internal_id"`
-	Title string `json:"title" db:"title"`
-	Description string `json:"description" db:"description"`
-	DueDate *time.Time `json:"due_date,omitempty" db:"due_date"`
-	Position int64 `json:"position" db:"position"`
-	CreatedAt time.Time `json:"created_at" db:"created_at"`
+	InternalID  int64      `json:"internal_id" db:"internal_id" gorm:"primaryKey;autoIncrement"`
+	PublicID    uuid.UUID  `json:"public_id" db:"public_id"`
+	ListID      int64      `json:"list_internal_id" db:"list_internal_id" gorm:"column:list_internal_id"`
+	Title       string     `json:"title" db:"title"`
+	Description string     `json:"description" db:"description"`
+	DueDate     *time.Time `json:"due_date,omitempty" db:"due_date"`
+	Position    int        `json:"position" db:"position"`
+	CreatedAt   time.Time  `json:"created_at" db:"created_at"`
 
-	//relasi
-	Assignees []CardAssignee `json:"assigness,omitempty" gorm:"foreignKey:CardID;references:InternalID"`
+	//relasi Assignees
+	Assignees   []CardAssignee   `json:"assigness,omitempty" gorm:"foreignKey:CardID;references:InternalID"`
 	Attachments []CardAttachment `json:"attachments,omitempty" gorm:"foreignKey:CardID;references:InternalID"`
-	Labels []CardLabel `json:"labels,omitempty" gorm:"foreignKey:CardID;references:InternalID"`
+	Labels      []CardLabel      `json:"labels,omitempty" gorm:"foreignKey:CardID;references:InternalID"`
 }
